gen_blank/requirements.txt
python-docx==1.1.2
docxcompose==1.4.0
Flask==3.0.3
click==8.1.7
python-dotenv=1.0.0
